requirements.txt
inquirer==3.4.0
openai==1.40.8
python-dotenv==1.0.1
tree_sitter_languages==1.10.2
